Moved all general functions to Utils
