+++ src/main/java/com/amaze/filemanager/ui/dialogs/SmbConnectDialog.java
@@ -28,6 +28,7 @@ import com.amaze.filemanager.utils.EditTextColorStateUtil;
 import com.amaze.filemanager.utils.Futils;
 import com.amaze.filemanager.utils.OpenMode;
 import com.amaze.filemanager.utils.PreferenceUtils;
+import com.amaze.filemanager.utils.Utils;
 import com.amaze.filemanager.utils.color.ColorUsage;
 import com.amaze.filemanager.utils.provider.UtilitiesProviderInterface;
 
@@ -198,7 +199,7 @@ public class SmbConnectDialog extends DialogFragment {
         EditTextColorStateUtil.setTint(user, accentColor);
         EditTextColorStateUtil.setTint(pass, accentColor);
 
-        Futils.setTint(ch, accentColor);
+        Utils.setTint(ch, accentColor);
         help.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
