+++ src/main/java/com/amaze/filemanager/utils/Utils.java
@@ -1,17 +1,112 @@
 package com.amaze.filemanager.utils;
 
 import android.content.Context;
+import android.content.res.ColorStateList;
+import android.graphics.Color;
+import android.graphics.drawable.Drawable;
 import android.os.Build;
 import android.support.annotation.ColorRes;
+import android.support.v4.content.ContextCompat;
+import android.support.v4.graphics.drawable.DrawableCompat;
+import android.view.View;
+import android.widget.CheckBox;
+
+import com.amaze.filemanager.R;
+
+import java.io.File;
+import java.text.SimpleDateFormat;
 
 /**
- * Contains useful functions and methods
+ * Contains useful functions and methods (NOTHING HERE DEALS WITH FILES)
  *
  * @author Emmanuel
  *         on 14/5/2017, at 14:39.
  */
 
 public class Utils {
+
+    private static final SimpleDateFormat sSDF = new SimpleDateFormat("MMM dd, yyyy");
+
+    //methods for fastscroller
+    /**
+     * TODO
+     * @param min
+     * @param max
+     * @param value
+     * @return
+     */
+    public static float getValueInRange(float min, float max, float value) {
+        float minimum = Math.max(min, value);
+        return Math.min(minimum, max);
+    }
+
+    /**
+     * TODO
+     * @param view
+     * @return
+     */
+    public static float getViewRawY(View view) {
+        int[] location = new int[2];
+        location[0] = 0;
+        location[1] = (int) view.getY();
+        ((View)view.getParent()).getLocationInWindow(location);
+        return location[1];
+    }
+
+    public static void setTint(CheckBox box, int color) {
+        if(Build.VERSION.SDK_INT>=21)return;
+        ColorStateList sl = new ColorStateList(new int[][]{
+                new int[]{-android.R.attr.state_checked},
+                new int[]{android.R.attr.state_checked}
+        }, new int[]{
+                Color.parseColor("#666666"),
+                color
+        });
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+            box.setButtonTintList(sl);
+        } else {
+            Drawable drawable = DrawableCompat.wrap(ContextCompat.getDrawable(box.getContext(), R.drawable.abc_btn_check_material));
+            DrawableCompat.setTintList(drawable, sl);
+            box.setButtonDrawable(drawable);
+        }
+    }
+
+    /**
+     * TODO
+     * @param f
+     * @return
+     */
+    public String getDate(File f) {
+        return getDate(f.lastModified());
+    }
+
+    /**
+     * TODO
+     * @param f
+     * @return
+     */
+    public static String getDate(long f) {
+        SimpleDateFormat sdf = new SimpleDateFormat("MMM dd yyyy | KK:mm a");
+        return sdf.format(f);
+    }
+
+
+    /**
+     * TODO
+     * @param f
+     * @param year
+     * @return
+     */
+    public static String getDate(long f, String year) {
+        String date = sSDF.format(f);
+        if(date.substring(date.length()-4,date.length()).equals(year))
+            date=date.substring(0,date.length()-6);
+        return date;
+    }
+
+    public boolean isAtleastKitkat(){
+        return Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT;
+    }
 
     /**
      * Gets color
